Made it faster ~ d:^)3
Added my own buffered reader to lose readAllFiles overhead.
Changed HashMaps to ArrayList(List) to avoid creating a lot of
Integer-Objects.
Removed possible boolean and just exit the program there.
Removed a lot of finals because it doesn't make a noticeable difference
in speed, but it does look strange.
Converted a stream into a for loop in an often called Method, which
speeded the program up by a bit.